requirements.txt
torch
transformers
torchvision
pillow
fastapi
uvicorn[standard]
python-multipart
python-dotenv
